VFB-318 Rebase fixes
