+++ src/app/parcels/batch/displayComponents/BatchParcelDataGrid.tsx
@@ -1,13 +1,18 @@
 "use client";
+
+import { Button, Dialog, DialogActions, DialogContent, DialogTitle } from "@mui/material";
 import { DataGrid, GridColDef } from "@mui/x-data-grid";
 import React, { useMemo, useState } from "react";
-import { useLocalStorage } from "@/app/parcels/batch/useLocalStorage";
+import { useLocalStorage, writeLocalTableState } from "@/app/parcels/batch/useLocalStorage";
 import { BatchTableDataState } from "@/app/parcels/batch/batchTypes";
 import batchParcelsReducer from "@/app/parcels/batch/batchParcelsReducer";
 import { tableStateToBatchDisplayRows } from "@/app/parcels/batch/displayHelpers";
 import { emptyBatchEditData, emptyOverrideData } from "@/app/parcels/batch/emptyData";
 import { DefaultTheme, useTheme } from "styled-components";
 import getCenteredBatchGridDisplayColumns from "@/app/parcels/batch/getCenteredBatchGridDisplayColumns";
+import { useRouter } from "next/navigation";
+import submitBatchTableData, { AddBatchRowError, displayUnsubmittedRows } from "@/app/parcels/batch/submitTableData";
+import { batchSubmitTestData, mockTableDataState } from "../mockData";
 export interface BatchGridDisplayRow {
     [key: string]: string | number | null;
     id: number;
@@ -51,9 +56,17 @@ export const defaultTableState: BatchTableDataState = {
 
 const BatchParcelDataGrid: React.FC = () => {
     const [tableState, dispatch] = useLocalStorage(batchParcelsReducer, defaultTableState);
+    const [dialogVisible, setDialogVisible] = useState(false);
+    const [dialogMessage, setDialogMessage] = useState("");
+    const [submitErrors, setSubmitErrors] = useState<AddBatchRowError[]>([]);
+    const router = useRouter();
+
+    writeLocalTableState(batchSubmitTestData);
+
     const displayRows = useMemo(() => {
         return tableStateToBatchDisplayRows(tableState);
     }, [tableState]);
+
     const theme: DefaultTheme = useTheme();
     const [isRowCollection, setIsRowCollection] = useState<{ [key: number]: boolean }>({});
     const centeredBatchGridDisplayColumns: GridColDef[] = getCenteredBatchGridDisplayColumns(
@@ -62,6 +75,42 @@ const BatchParcelDataGrid: React.FC = () => {
         isRowCollection,
         setIsRowCollection
     );
+
+    const handleSubmit = async (): Promise<void> => {
+        console.log(tableState.batchDataRows)
+        const { errors: submitErrors } = await submitBatchTableData(tableState);
+        console.log(submitErrors)
+        setSubmitErrors(submitErrors);
+
+        if (submitErrors.length == 0) {
+            setDialogMessage(`Successfully submitted ${tableState.batchDataRows.length} row(s)`);
+            handleReset();
+        } else {
+            setDialogMessage(
+                `There were ${submitErrors.length} error(s) encountered during submission.`
+            );
+            handleUnsubmittedRows(submitErrors);
+        }
+
+        setDialogVisible(true);
+    };
+
+    const handleReset = (): void => {
+        dispatch({
+            type: "initialise_table_state",
+            payload: { initialTableState: defaultTableState },
+        });
+    };
+
+    const handleUnsubmittedRows = (submitErrors: AddBatchRowError[]): void => {
+        const newState = displayUnsubmittedRows(tableState, submitErrors);
+        dispatch({ type: "initialise_table_state", payload: { initialTableState: newState } });
+    };
+
+    const handleReturnToParcels = (): void => {
+        router.push("/parcels");
+    };
+
     return (
         <>
             <DataGrid
@@ -121,6 +170,7 @@ const BatchParcelDataGrid: React.FC = () => {
                     },
                     "& .Mui-error": {
                         backgroundColor: `${theme.error}`,
+                        color: `${theme.text}`,
                     },
                     border: "1px solid",
                     borderColor: `${theme.main.border}`,
@@ -128,6 +178,83 @@ const BatchParcelDataGrid: React.FC = () => {
                 }}
                 hideFooter
             />
+            <Button
+                onClick={handleSubmit}
+                variant="contained"
+                sx={{ marginLeft: "1rem", minWidth: "120px" }}
+            >
+                Submit
+            </Button>
+            <Dialog
+                open={dialogVisible}
+                sx={{
+                    "& .MuiDialog-paper": {
+                        padding: "1rem",
+                        borderRadius: "15px",
+                        boxShadow: "0 3px 5px rgba(0,0,0,0.8)",
+                        margin: "auto",
+                        borderColor: submitErrors.length == 0 ? "#a8d49c" : "#990f0f",
+                        borderWidth: "1px",
+                        borderStyle: "solid",
+                    },
+                }}
+            >
+                <DialogTitle
+                    sx={{
+                        fontSize: "1.2rem",
+                        textAlign: "center",
+                    }}
+                >
+                    {dialogMessage}
+                </DialogTitle>
+                {submitErrors.length !== 0 && (
+                    <DialogContent sx={{ textAlign: "center" }}>
+                        An unexpected error occurred and some rows did not submit.
+                        <br />
+                        Please try submitting again.
+                    </DialogContent>
+                )}
+                <DialogActions sx={{ justifyContent: "center" }}>
+                    {submitErrors.length > 0 ? (
+                        <>
+                            <Button
+                                variant="contained"
+                                sx={{ margin: "1rem", minWidth: "120px" }}
+                                onClick={() => {
+                                    handleReset();
+                                    setDialogVisible(false);
+                                }}
+                            >
+                                Reset rows and start again
+                            </Button>
+                            <Button
+                                variant="contained"
+                                onClick={() => setDialogVisible(false)}
+                                sx={{ margin: "1rem", minWidth: "120px" }}
+                            >
+                                View unsubmitted rows
+                            </Button>
+                        </>
+                    ) : (
+                        <>
+                            <Button
+                                variant="contained"
+                                sx={{ margin: "1rem", minWidth: "120px" }}
+                                onClick={() => setDialogVisible(false)}
+                            >
+                                Add more rows
+                            </Button>
+                            <Button
+                                variant="contained"
+                                sx={{ margin: "1rem", minWidth: "120px" }}
+                                onClick={handleReturnToParcels}
+                            >
+                                Return to the Parcels Page
+                            </Button>
+                        </>
+                    )}
+                </DialogActions>
+            </Dialog>
         </>
     );
 };
